requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
flake8==7.0.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-MySQL==1.5.2
Flask-Restless==0.17.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
mccabe==0.7.0
mimerender==0.6.0
mistune==3.0.2
mysql-connector-python==8.4.0
mysqlclient==2.2.4
packaging==24.0
pluggy==1.5.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyMySQL==1.1.1
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-mimeparse==1.6.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
SQLAlchemy==2.0.30
testing.common.database==2.0.3
typing_extensions==4.11.0
Werkzeug==3.0.3
